feat: update admin panel, create custom hook and modal

+++ frontend/src/pages/admin/AdminFeedback.tsx
@@ -1,12 +1,47 @@
 import { Link } from "react-router-dom";
 import useCotactMessageStore from "../../store/contactMessageStore";
+import { toast } from "react-toastify";
+import { ModalDeleteAll } from "../../components/Modal/Modal";
+import { useState } from "react";
 
 const AdminFeedback = () => {
-    const { contactMessage, deleteContactMessage } = useCotactMessageStore()
+    const { contactMessage, deleteContactMessage, deleteAllContactMessage, fetchContactsMessages } = useCotactMessageStore()
+    const [modalActive, setModalActive] = useState<boolean>(false)
+
+    // Обработка модалки
+    const handleDeleteAll = async () => {
+        try {
+            await deleteAllContactMessage();
+            toast.success('Все сообщения удалены');
+        } catch (error) {
+            toast.error('Ошибка при удалении');
+        }
+    };
 
     return(
         <div className="w-full">
-            <h1 className="text-3xl text-white mb-5">Обратная связь</h1>
+            
+            <div className="flex mb-5 gap-3">
+                <h1 className="text-3xl text-white mr-5">Обратная связь</h1>
+                <button
+                    onClick={() => {
+                        toast.success('Обновлен')
+                        fetchContactsMessages()
+                    }}
+                    className="block text-center bg-gray-600 text-white px-4 py-2 rounded hover:bg-gray-700 
+                                duration-300 font-medium text-sm transition-all"
+                >
+                    Обновить
+                </button>
+
+                <button
+                    onClick={() => setModalActive(true)}
+                    className="block text-center bg-red-600 text-white px-4 py-2 rounded hover:bg-red-700 
+                                duration-300 font-medium text-sm transition-all"
+                >
+                    Очистить
+                </button>
+            </div>
 
             <div className="grid grid-cols-1 sm:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4 gap-5">
                 {contactMessage.map((message) => (
@@ -15,7 +50,6 @@ const AdminFeedback = () => {
                         className="bg-gray-800 rounded-lg p-5 duration-1000 transform 
                         hover:shadow-xl border border-gray-700 flex flex-col h-full hover:bg-gray-900"
                     >
-                        {/* Заголовок с иконкой */}
                         <div className="flex items-start justify-between mb-3">
                             <div className="flex items-center gap-2">
                                 <span className="text-lg text-white">📧 Тема: </span>
@@ -24,11 +58,10 @@ const AdminFeedback = () => {
                                 </h3>
                             </div>
                             <span className="text-xs text-gray-500 bg-gray-900 px-2 py-1 rounded">
-                                # {message._id?.slice(-6) || '000000'}
+                                # {message._id?.slice(-5) || '00000'}
                             </span>
                         </div>
 
-                        {/* Информация об отправителе */}
                         <div className="space-y-2 flex-1">
                             <div className="flex items-center gap-2 text-gray-300">
                                 <span className="text-sm">👤</span>
@@ -44,24 +77,23 @@ const AdminFeedback = () => {
                             </div>
                         </div>
 
-                        {/* Кнопка */}
                         <div className="mt-4 pt-3 gap-3 flex border-gray-600">
                             <Link
                                 to={`/contacts/${message._id}`}
                                 className="w-6/12 block text-center bg-blue-600 text-white px-4 py-2 rounded 
-                                    hover:bg-blue-700 duration-300 font-medium text-sm
-                                    transition-all"
+                                    hover:bg-blue-700 duration-300 font-medium text-sm transition-all"
                             >
                                 Ознакомится
                             </Link>
 
                             <button
                                 className="w-6/12 block text-center bg-red-600 text-white px-4 py-2 rounded 
-                                    hover:bg-red-700 duration-300 font-medium text-sm
-                                    transition-all"
-                                    onClick={() => {
-                                        deleteContactMessage(message._id)
-                                    }}
+                                    hover:bg-red-700 duration-300 font-medium text-sm transition-all"
+                                onClick={() => {
+                                    if (window.confirm('Удалить это сообщение?')) {
+                                        deleteContactMessage(message._id);
+                                    }
+                                }}
                             >
                                 Удалить
                             </button>
@@ -69,6 +101,13 @@ const AdminFeedback = () => {
                     </div>
                 ))}
             </div>
+
+            {/* Модалка */}
+            <ModalDeleteAll 
+                active={modalActive} 
+                setActive={setModalActive}
+                onConfirm={handleDeleteAll}
+            />
         </div>
     )
 }
